Create king_crab_code.ino
Began the main code file. Made basic functions for the state machine in terms of calculating the netPoints inside the claws and assigned placeholder functions for what will occur dependant on netPoints (discard or keep the balls).

Similarly, a function will be produced for when to tilt the camera to begin calculating the netPoints.